+++ CHANGELOG.md
@@ -8,3 +8,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 - Integration with FAST (Fast Audio Server for Testing), allowing us to test compatibility on any platform *before* hooking up that platform's audio API(s).
+- `ui_timecode_ms` setting. When set to `true`, MPL displays milliseconds in timecodes (i.e track position, track duration).
+- Support for boolean settings.